archive/requirements.txt
pandas
scikit-learn
mysql-connector-python
pymysql
joblib
requests
matplotlib
seaborn
